Separate DataSink config dataclasses from implementations
Learned from internal schema/v2alpha1/data_sink.py where HiveSink(JSONData)
is a typed config object. Applied the same separation in the OS package:

- schema/data_sink.py: now only DataSink ABC + SinkResult (contracts)
- workflow/schema/sinks/: typed config dataclasses (HiveSinkConfig,
  LocalFileSinkConfig, InMemorySinkConfig) validated at pipeline-definition time
- workflow/sinks/: implementations taking typed configs (HiveSink, LocalFileSink,
  InMemorySink), shared across tasks (not nested under any task subfolder)
- DatasetPluginConfig.__post_init__ uses LocalFileSinkConfig + LocalFileSink
  from new locations for the destination_path shorthand
