+++ python/michelangelo/workflow/schema/sinks/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for michelangelo.workflow.schema.sinks."""
