+++ python/michelangelo/workflow/sinks/__init__.py
@@ -1,0 +1,22 @@
+"""Built-in DataSink implementations for DatasetPusherPlugin.
+
+Each sink accepts a typed config dataclass from
+``michelangelo.workflow.schema.sinks`` — validated at pipeline-definition time:
+
+    from michelangelo.workflow.schema.sinks import HiveSinkConfig, LocalFileSinkConfig
+    from michelangelo.workflow.sinks import HiveSink, LocalFileSink, InMemorySink
+
+    sinks = [
+        LocalFileSink(LocalFileSinkConfig("/tmp/out")),
+        HiveSink(HiveSinkConfig(database="ml", table="predictions")),
+    ]
+
+Provider layers extend this by subclassing ``DataSink`` and registering their
+own config dataclasses in ``schema/sinks/``.
+"""
+
+from michelangelo.workflow.sinks.hive import HiveSink
+from michelangelo.workflow.sinks.local import LocalFileSink
+from michelangelo.workflow.sinks.memory import InMemorySink
+
+__all__ = ["HiveSink", "InMemorySink", "LocalFileSink"]
