+++ python/michelangelo/workflow/sinks/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for michelangelo.workflow.sinks."""
